C variadic functions - Advance

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -22,11 +22,11 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 
 	for (i = 0; i < (n - 1); i++)
 		if (!separator)
-			printf("%d", va_arg(args, int));
+			printf("%ld", va_arg(args, long));
 		else
-			printf("%d%s", va_arg(args, int), separator);
+			printf("%ld%s", va_arg(args, long), separator);
 
-	printf("%d\n", va_arg(args, int));
+	printf("%ld\n", va_arg(args, long));
 
 	va_end(args);
 }
